src/requirements.txt
discord-timers==0.1.0
beautifulsoup4
aiohttp>=3.8.0
discord.py>=2.5.2
python-dotenv
requests
Pillow  # Aktualisiert für Meme-Generator
psutil  # Für Systemstatistiken